Tmux integration fixes for tmux1.8

+++ GuiTerminalWindow.h
@@ -105,7 +105,7 @@ public:
 
     void keyPressEvent ( QKeyEvent * e );
     void keyReleaseEvent ( QKeyEvent * e );
-    int from_backend(int is_stderr, const char *data, int len);
+    int from_backend(int is_stderr, const char *data, size_t len);
     void preDrawTerm();
     void drawTerm();
     void drawText(int row, int col, wchar_t *ch, int len, unsigned long attr, int lattr);
